test/receiver-py/requirements.txt
azure-eventhub
python-dotenv
